% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbh-package.R
\docType{package}
\name{bbh-package}
\alias{bbh-package}
\alias{bbh}
\alias{bbhsst}
\title{bbh package}
\description{
Serving Boothbay Harbor SST.
}
\section{data source}{

Serves up DMR's data for BBH \href{https://dmr-maine.opendata.arcgis.com/datasets/maine::mainedmr-boothbay-harbor-sea-surface-temperatures/about}{bbh sst}
}

