% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bind_rows_sst}
\alias{bind_rows_sst}
\title{Binds rows of SST data in a list}
\usage{
bind_rows_sst(x)
}
\arguments{
\item{x}{a list of one or more sst class tibbles}
}
\value{
a tibble of class sst_group
}
\description{
Binds rows of SST data in a list
}
