% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{E01_bydate}
\alias{E01_bydate}
\title{Create by-date E01 data}
\usage{
E01_bydate(x = fetch_E01())
}
\arguments{
\item{x}{character, the E01 dataset}
}
\value{
tibble
}
\description{
Create by-date E01 data
}
