% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{get_path}
\alias{get_path}
\title{Sets path name}
\usage{
get_path(..., root = "C:\\\\Users\\\\ernst\\\\AppData\\\\Local/bbh/bbh")
}
\arguments{
\item{...}{any filepath segments, optional}

\item{root}{root filepath, defaults to user directory}
}
\value{

}
\description{
Sets path name
}
