% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sst}
\alias{plot.sst}
\title{Plots SST data}
\usage{
\method{plot}{sst}(x, xlim = range(x$COLLECTION_DATE, na.rm = TRUE), ...)
}
\arguments{
\item{x}{sst class tibble}

\item{xlim}{date range for plotting}

\item{...}{future other arguments passed to ggplot2}
}
\value{
ggplot2 object
}
\description{
Plots SST data
}
