% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sst_group}
\alias{plot.sst_group}
\title{Plots SST group data from rbind.sst}
\usage{
\method{plot}{sst_group}(x, ...)
}
\arguments{
\item{x}{sst class tibble}

\item{...}{potential future other arguments passed to ggplot2}
}
\value{
ggplot2 object
}
\description{
Plots SST group data from rbind.sst
}
