% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rbind.sst}
\alias{rbind.sst}
\title{Binds rows of SST data}
\usage{
\method{rbind}{sst}(...)
}
\arguments{
\item{...}{one of more class sst tibbles passed to base rbind}
}
\value{
a tibble of class sst_group
}
\description{
Binds rows of SST data
}
